requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
celery==5.6.0
redis==7.1.0
Pillow==12.0.0
django-storages==1.14.2
boto3==1.28.45
APScheduler==3.10.4
openpyxl==3.1.5
reportlab==4.0.7
pandas==2.0.3
